Refactor combat engine UI separation
